Add conditional imports
